+++ src/app/bank-account/_shared/components/bank-account-card/bank-account-card.tsx
@@ -11,9 +11,9 @@ import Dropdown from '@/shared/components/ui/dropdown/dropdown';
 import { useModalStore } from '@/shared/components/ui/modal/modal-store';
 import { toastify } from '@/shared/components/ui/toast/toastify';
 import { useNavigation } from '@/shared/hooks/navigation/use-navigation';
-import { handleError } from '@/shared/hooks/use-error';
 import { BankCode } from '@/shared/types';
 import { getBankIconByCode, getBankNameByCode } from '@/shared/utils/bank';
+import { handleError } from '@/shared/utils/error';
 
 import styles from './bank-account-card.module.scss';
